requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiolimiter==1.1.0
aioredis==2.0.1
aiosignal==1.3.1
annotated-types==0.7.0
anthropic==0.49.0
anyio==4.9.0
asttokens==2.4.1
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.1.0
backoff==2.2.1
beautifulsoup4==4.12.0
blinker==1.8.2
BM25==1.0.0
cachetools==5.5.2
certifi==2023.7.22
chardet==5.2.0
charset-normalizer==3.3.0
click==8.1.7
cloudpickle==2.2.1
comm==0.2.1
contourpy==1.2.0
cycler==0.12.1
datasets==2.14.5
debugpy==1.8.0
decorator==5.1.1
dill==0.3.7
distro==1.9.0
eval_type_backport==0.2.2
exceptiongroup==1.2.0
executing==2.0.1
filelock==3.18.0
Flask==3.0.3
Flask-Cors==5.0.0
fonttools==4.47.0
frozenlist==1.4.0
fsspec==2023.6.0
ghdiff==0.4
google-api-core==2.24.2
google-api-python-client==2.166.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-genai==1.9.0
googleapis-common-protos==1.69.2
gunicorn==20.1.0
gym==0.26.2
gym-notices==0.0.8
h11==0.14.0
html2text==2020.1.16
httpcore==1.0.2
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.17.3
idna==3.4
ipykernel==6.28.0
ipython==8.19.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jiter==0.9.0
jupyter_client==8.6.0
jupyter_core==5.7.0
kiwisolver==1.4.5
linecache2==1.0.0
Markdown==3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.2
matplotlib-inline==0.1.6
mdurl==0.1.2
multidict==6.0.4
multiprocess==0.70.15
nest-asyncio==1.5.8
numpy==1.26.0
oauthlib==3.2.2
openai==1.71.0
openlimit==0.3.0
packaging==23.2
pandas==2.1.1
parse==1.19.1
parso==0.8.3
pexpect==4.9.0
pickleshare==0.7.5
pillow==11.1.0
platformdirs==4.1.0
prompt-toolkit==3.0.43
propcache==0.3.1
proto-plus==1.26.1
protobuf==6.30.2
psutil==5.9.7
ptvsd==4.3.2
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==13.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.2
pydantic_core==2.33.1
Pygments==2.17.2
pyparsing==3.1.1
PyPDF2==3.0.1
pyperclip==1.8.2
python-dateutil==2.8.2
python-dotenv==1.0.0
python-leetcode==1.2.1
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.2
rank-bm25==0.2.2
regex==2023.10.3
requests==2.31.0
requests-oauthlib==2.0.0
rich==13.9.4
rsa==4.9
scipy==1.11.4
seaborn==0.13.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
sseclient-py==1.8.0
stack-data==0.6.3
tabulate==0.9.0
tiktoken==0.5.1
together==1.5.5
tokenizers==0.15.2
tornado==6.4
tqdm==4.67.1
traceback2==1.4.0
traitlets==5.14.0
typer==0.15.2
typing-inspection==0.4.0
typing_extensions==4.13.1
tzdata==2023.3
unittest2==1.1.0
uritemplate==4.1.1
urllib3==2.0.6
wcwidth==0.2.12
websockets==15.0.1
Werkzeug==3.0.4
xxhash==3.4.1
yarl==1.19.0
